+++ griptape/drivers/web_search/tavily_web_search_driver.py
@@ -1,11 +1,10 @@
 from __future__ import annotations
 
-import json
 from typing import TYPE_CHECKING, Any
 
 from attrs import Factory, define, field
 
-from griptape.artifacts import ListArtifact, TextArtifact
+from griptape.artifacts import JsonArtifact, ListArtifact
 from griptape.drivers import BaseWebSearchDriver
 from griptape.utils import import_optional_dependency
 
@@ -25,9 +24,4 @@ class TavilyWebSearchDriver(BaseWebSearchDriver):
     def search(self, query: str, **kwargs) -> ListArtifact:
         response = self.client.search(query, max_results=self.results_count, **self.params, **kwargs)
         results = response["results"]
-        return ListArtifact(
-            [
-                TextArtifact(json.dumps({"title": result["title"], "url": result["url"], "content": result["content"]}))
-                for result in results
-            ]
-        )
+        return ListArtifact([JsonArtifact(result, encoding="utf-8") for result in results])